Cache improved, codebase improved
