+++ awt-admin/jobs/pages.php
@@ -8,7 +8,7 @@ include_once JOBS . 'loaders' . DIRECTORY_SEPARATOR . 'awt-pluginLoader.php';
 include_once JOBS . 'awt-domainBuilder.php';
 include_once JOBS . 'loaders' . DIRECTORY_SEPARATOR . 'awt-themesLoader.php';
 
-use paging\paging;
+use paging\{paging, editor};
 use admin\{authentication, profiler};
 
 $check = new authentication;
@@ -20,10 +20,12 @@ if (!$check->checkAuthentication()) {
 }
 
 $pages = new paging(array());
+$editor = new editor(array());
+
 if(isset($_POST['getPages'])) echo json_encode($pages->getAllPages());
-if(isset($_POST['getAllPages'])) echo json_encode($pages->getEveryPage());
-if(isset($_POST['createEmpty']) && isset($_POST['name'])) die(json_encode($pages->createEmptyPage($_POST['name'])));
+if(isset($_POST['getAllPages'])) echo json_encode($editor->getEveryPage());
+if(isset($_POST['createEmpty']) && isset($_POST['name'])) die(json_encode($editor->createEmptyPage($_POST['name'])));
 if(isset($_POST['deletePage']) && isset($_POST['id']) && $profiler->checkPermissions(0)) echo json_encode($pages->deletePage($_POST['id']));
 if(isset($_POST['update']) && isset($_POST['change']) && isset($_POST['value']) && $profiler->checkPermissions(0)) {
-    echo json_encode($pages->changeInfo($_POST['update'], $_POST['change'], $_POST['value']));
+    echo json_encode($editor->changeInfo($_POST['update'], $_POST['change'], $_POST['value']));
 }
