+++ awt-src/classes/paging/paging.class.php
@@ -5,7 +5,6 @@ namespace paging;
 use admin\profiler;
 use cache\cache;
 use database\databaseConfig;
-use notifications\notifications;
 
 class paging extends cache
 {
@@ -45,6 +44,25 @@ class paging extends cache
         $this->mysqli = $this->database->getConfig();
     }
 
+    public function searchPage(string $column, string $value) : bool
+    {
+        $result = array();
+        $stmt = $this->mysqli->prepare("SELECT * FROM `awt_paging` WHERE `" . $column . "` = ?;");
+        $stmt->bind_param('s', $value);
+        $stmt->execute();
+        $stmt->store_result();
+        $stmt->bind_result($result['id'], $result['name'], $result['description'], $result['content_1'], $result['content_2'], $result['status'], $result['token'], $result['override']);
+        $stmt->fetch();
+
+        if ($stmt->num_rows == 1) {
+            $stmt->close();
+            return true;
+        }
+
+        return false;
+    }
+
+
     public function addBuiltInPage(string $name, string $path, string $description = '')
     {
         global $builtInPages;
@@ -53,7 +71,7 @@ class paging extends cache
         $builtInPages[] = array('name'=> $name, 'builtIn' => true);
     }
 
-    public function getPage(bool $selfCalled = false, string $varName = '')
+    public function getPage(bool $loadAdmin = false, bool $selfCalled = false, string $varName = '') : void
     {
         global $theme;
         global $menu;
@@ -64,7 +82,7 @@ class paging extends cache
 
         if ($this->cacheEnabled && $this->checkForCache($_GET['page'])) {
             echo $this->readCache($_GET['page']);
-            return 1;
+            return;
         }
 
         $result = array();
@@ -80,39 +98,42 @@ class paging extends cache
 
         if ($stmt->num_rows == 1) {
 
-            echo $result['content_1'] . $result['content_2'];
             if ($this->cacheEnabled)
                 $this->writePageCache($_GET['page'], $result['content_1'] . $result['content_2']);
-
+        
             $stmt->close();
-            return 1;
+            echo $result['content_1'] . $result['content_2'];
+            return;
+
+        } else {
+            $stmt->close();
         }
 
-        $stmt->close();
+
 
         if ($selfCalled && $varName != '') {
             global $$varName;
             $$varName = $this;
         }
 
-        if (isset($_GET['page'])) {
+        if (isset($_GET['page']) && $loadAdmin) {
             if (array_key_exists($_GET['page'], $this->adminPages)) {
                 if (file_exists($this->adminPages[$_GET['page']])) {
                     include_once $this->adminPages[$_GET['page']];
-                    return 1;
+                    return;
                 }
             }
 
             if (array_key_exists($_GET['page'], $this->pluginPages)) {
                 if (file_exists($this->pluginPages[$_GET['page']])) {
                     include_once $this->pluginPages[$_GET['page']];
-                    return 1;
+                    return;
                 }
             }
             if (array_key_exists($_GET['page'], $this->pages)) {
                 if (file_exists($this->pages[$_GET['page']]['path'])) {
                     include_once $this->pages[$_GET['page']]['path'];
-                    return 1;
+                    return;
                 }
             }
 
@@ -120,81 +141,7 @@ class paging extends cache
         }
     }
 
-
-    public function uploadPage(string $name, string $page, string $status = "live", int $override = 0)
-    {
-
-        $result = array();
-        $stmt = $this->mysqli->prepare("SELECT * FROM `awt_paging` WHERE `name` = ?;");
-        $stmt->bind_param('s', $name);
-        $stmt->execute();
-        $stmt->store_result();
-        $stmt->bind_result($result['id'], $result['name'], $result['content_1'], $result['content_2'], $result['status'], $result['token'], $result['override']);
-        $stmt->fetch();
-
-        if ($stmt->num_rows == 1) {
-
-            $this->updatePage($name, $page, $status, $override);
-
-            $stmt->close();
-            return 1;
-        }
-
-        $page_lenght = strlen($page);
-
-        if ($page_lenght > 16777215) {
-            $content_1 = substr($page, 0, 16777214);
-            $content_2 = substr($page, 16777214, 16777214 * 2);
-        } else {
-            $content_1 = $page;
-            $content_2 = "";
-        }
-
-        $token = hash("sha512", time());
-
-        $stmt = $this->mysqli->prepare("INSERT INTO `awt_paging`(`id`, `name`, `content_1`, `content_2`, `status`, `token`, `override`) VALUES (NULL, ?, ?, ?, ?, ?, ?)");
-
-        $stmt->bind_param("sssssi", $name, $content_1, $content_2, $status, $token, $override);
-        $stmt->execute();
-        $stmt->close();
-
-        $profiler = new profiler();
-
-        $notifications = new notifications("Pages", $profiler->name . " has uploaded content to $name page.");
-        $notifications->pushNotification();
-    }
-
-    public function updatePage(string $name, string $page, string $status = "live", int $override = 0)
-    {
-
-        $page_length = strlen($page);
-
-        $page = str_replace("ui-sortable-handle", "", $page);
-        $page = str_replace("ui-sortable", "", $page);
-        $page = str_replace("ui-sortable", "", $page);
-        $page = str_replace('contenteditable="true"', "", $page);
-        $page = str_replace('selected', "", $page);
-
-        if ($page_length > 16777215) {
-            $content_1 = substr($page, 0, 16777214);
-            $content_2 = substr($page, 16777214, 16777214 * 2);
-        } else {
-            $content_1 = $page;
-            $content_2 = "";
-        }
-
-        $stmt = $this->mysqli->prepare("UPDATE `awt_paging` SET `content_1` = ?, `content_2` = ?, `status` = ?, `override` = ? WHERE `name` = ?");
-        $stmt->bind_param("sssis", $content_1, $content_2, $status, $override, $name);
-        $stmt->execute();
-        $stmt->close();
-
-        $profiler = new profiler();
-
-        $notifications = new notifications("Pages", $profiler->name . " has updated content of $name page.");
-        $notifications->pushNotification();
-    }
-
-    public function getAllPages()
+    public function getAllPages() : array
     {
         $result = array();
 
@@ -206,101 +153,7 @@ class paging extends cache
             $result[] = $row;
         }
 
-
         return $result;
-    }
-
-    public function createEmptyPage(string $name)
-    {
-        $content_1 = "<div class='pageSection'>";
-        $content_2 = "</div>";
-        $status = "preview";
-        $override = 0;
-
-        $token = hash("sha512", time());
-
-        $stmt = $this->mysqli->prepare("INSERT INTO `awt_paging`(`id`, `name`, `content_1`, `content_2`, `status`, `token`, `override`) VALUES (NULL, ?, ?, ?, ?, ?, ?)");
-
-        $stmt->bind_param("sssssi", $name, $content_1, $content_2, $status, $token, $override);
-        $stmt->execute();
-        $stmt->close();
-
-        $profiler = new profiler();
-
-        $notifications = new notifications("Pages", $profiler->name . " has created new page: $name.");
-        $notifications->pushNotification();
-
-        return "OK";
-    }
-
-    public function deletePage(int $id)
-    {
-        $stmt = $this->mysqli->prepare("DELETE FROM `awt_paging` WHERE `id` = ?;");
-        $stmt->bind_param("i", $id);
-        $stmt->execute();
-        $stmt->close();
-
-        $profiler = new profiler();
-
-        $notifications = new notifications("Pages", $profiler->name . " has deleted page with ID: $id.", "high");
-        $notifications->pushNotification();
-
-        return "Page deleted with id: $id";
-    }
-
-    public function loadPageEdit(int $id)
-    {
-
-        $result = array();
-
-        $stmt = $this->mysqli->prepare("SELECT * FROM `awt_paging` WHERE `id` = ?;");
-        $stmt->bind_param('s', $id);
-        $stmt->execute();
-        $stmt->store_result();
-        $stmt->bind_result($result['id'], $result['description'], $result['name'], $result['content_1'], $result['content_2'], $result['status'], $result['token'], $result['override']);
-        $stmt->fetch();
-
-
-        if ($stmt->num_rows == 1) {
-
-            echo $result['content_1'] . $result['content_2'];
-            $stmt->close();
-            return 1;
-        } else {
-            die("Page does not exist");
-        }
-    }
-
-    public function loadPage($page, string $mode = "id")
-    {
-        $result = array();
-
-        $status = "live";
-
-        if ($mode == "id") {
-            $stmt = $this->mysqli->prepare("SELECT * FROM `awt_paging` WHERE `id` = ? AND `status` = ?;");
-            $stmt->bind_param('ss', $page, $status);
-            $stmt->execute();
-            $stmt->store_result();
-            $stmt->bind_result($result['id'], $result['description'], $result['name'], $result['content_1'], $result['content_2'], $result['status'], $result['token'], $result['override']);
-            $stmt->fetch();
-        } else if ($mode == "name") {
-            $stmt = $this->mysqli->prepare("SELECT * FROM `awt_paging` WHERE `name` = ? AND `status` = ?;");
-            $stmt->bind_param('ss', $page, $status);
-            $stmt->execute();
-            $stmt->store_result();
-            $stmt->bind_result($result['id'], $result['description'], $result['name'], $result['content_1'], $result['content_2'], $result['status'], $result['token'], $result['override']);
-            $stmt->fetch();
-        }
-
-        if ($stmt->num_rows == 1) {
-
-            echo $result['content_1'] . $result['content_2'];
-            $stmt->close();
-            return 1;
-        } else {
-            die("Page does not exist");
-        }
     }
 
     public function loadPreview(string $token, string $name)
@@ -324,25 +177,6 @@ class paging extends cache
         } else {
             die("Page does not exist");
         }
-    }
-
-    public function getEveryPage() : array
-    {   
-        global $builtInPages;
-        $result = $this->getAllPages();
-        $result = array_merge($builtInPages, $result);
-        return $result;
-    }
-
-    public function changeInfo(int $id, string $change, string $value) : bool
-    {
-        $stmt = $this->mysqli->prepare("UPDATE `awt_paging` SET `" . $change . "` = ? WHERE `id` = ?");
-
-        $stmt->bind_param("si", $value, $id);
-
-        if($stmt->execute()) return true;
-
-        return false;
     }
 
 }
